refactor player controller into seperate monoBehavior scripts
