+++ Assets/Characters/Player/Monk/Prefabs/PlayerCharacter.prefab
@@ -307,6 +307,9 @@ GameObject:
   - component: {fileID: 511697649317852147}
   - component: {fileID: 3993709810264871192}
   - component: {fileID: 7496793136723985154}
+  - component: {fileID: -7000680862678952261}
+  - component: {fileID: -2524199032906416086}
+  - component: {fileID: 4392064198536165683}
   m_Layer: 0
   m_Name: PlayerCharacter
   m_TagString: Player
@@ -497,9 +500,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   AoeAttack: {fileID: 3804288888145762448, guid: 1edcbcedf99014440a1cb7c63bcf5f02, type: 3}
   SwordSwingAttack: {fileID: 1220795059866920382}
-  Score: {fileID: 0}
   Movement: {fileID: -440124971067648005, guid: 23bf02c5a505b7e469d359c6774e31c6, type: 3}
-  healthPoints: 100
 --- !u!114 &7496793136723985154
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -514,6 +515,45 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   moveSpeed: 7
   collisionOffset: 0.5
+--- !u!114 &-7000680862678952261
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8111309106597192690}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4fc12ec1ae0be074c91fb1d4a5fa81e3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  HealthGlobe: {fileID: 0}
+  healthPoints: 100
+--- !u!114 &-2524199032906416086
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8111309106597192690}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fa17d5b8dcfaac2448974691735db806, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  UiScore: {fileID: 0}
+--- !u!114 &4392064198536165683
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8111309106597192690}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 38dd48514c4c74645b81cec73922a35e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &3075538869245722340
 PrefabInstance:
   m_ObjectHideFlags: 0
