+++ Assets/Scripts/Characters/Enemies/Versions/MeleeEnemy.cs
@@ -19,7 +19,7 @@ namespace Assets.Scripts.Characters.Enemies.Versions
             if (AnimationIsPlaying(new string[] { "TakeHit", "Death", "Attack1", "Attack2" }))
                 return;
 
-            if (Vector3.Distance(playerController.transform.position, transform.position) > meleeAttackRange)
+            if (Vector3.Distance(playerStats.transform.position, transform.position) > meleeAttackRange)
                 return;
 
             var attack = RandomHelper.GetRandom(40) ? "Attack2" : "Attack1";
@@ -33,17 +33,17 @@ namespace Assets.Scripts.Characters.Enemies.Versions
             if (command == "Attack1")
             {
                 // Check if still in melee range
-                if (Vector3.Distance(playerController.transform.position, transform.position) < meleeAttackRange)
+                if (Vector3.Distance(playerStats.transform.position, transform.position) < meleeAttackRange)
                 {
-                    playerController.TakeDamage(Attack1Damage);
+                    playerStats.TakeDamage(Attack1Damage);
                 }
             }
             if (command == "Attack2")
             {
                 // Check if still in melee range
-                if (Vector3.Distance(playerController.transform.position, transform.position) < meleeAttackRange)
+                if (Vector3.Distance(playerStats.transform.position, transform.position) < meleeAttackRange)
                 {
-                    playerController.TakeDamage(Attack2Damage);
+                    playerStats.TakeDamage(Attack2Damage);
                 }
             }
         }
