+++ Assets/Scripts/Characters/Player/PlayerController.cs
@@ -1,12 +1,9 @@
-using Assets.Scripts.Characters.Enemies;
-using Assets.Scripts.Helpers;
-using Assets.Scripts.Items;
-using Assets.Scripts.UI;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
 namespace Assets.Scripts.Characters.Player
 {
+    [RequireComponent(typeof(Animator))]
     [RequireComponent(typeof(PlayerMovement))]
     public class PlayerController : MonoBehaviour
     {
@@ -14,44 +11,21 @@ namespace Assets.Scripts.Characters.Player
         public GameObject AoeAttack, SwordSwingAttack;
 
         [SerializeField]
-        public GameObject Score;
-
-        [SerializeField]
         private InputActionReference Movement;
 
-        [SerializeField]
-        private int healthPoints = 100;
-
-        private int currentHP;
-
         private PlayerMovement playerMovement;
-        private PlayerAttacks playerAttacks;
-
         private Animator animator;
-        private SpriteRenderer spriteRenderer;
-        private HealthGlobe healthGlobe;
+
+        private GameObject activeAoeAttack;
 
         private void OnEnable()
         {
-            var rb = GetComponent<Rigidbody2D>();
             animator = GetComponent<Animator>();
-            spriteRenderer = GetComponent<SpriteRenderer>();
             playerMovement = GetComponent<PlayerMovement>();
-            playerAttacks = new PlayerAttacks(animator);
-
-            currentHP = healthPoints;
-            healthGlobe = FindObjectOfType<HealthGlobe>();
-        }
-
-        public Vector3 GetSpriteCenter()
-        {
-            return spriteRenderer.bounds.center;
         }
 
         private void FixedUpdate()
         {
-            AggroMobs();
-            AggroItems();
             var value = Movement.action.ReadValue<Vector2>();
             playerMovement.Move(value);
         }
@@ -67,13 +41,19 @@ namespace Assets.Scripts.Characters.Player
         void OnMainAttack()
         {
             if (SwordSwingAttack != null)
-                playerAttacks.SwordSwing(SwordSwingAttack);
+            {
+                SwordSwingAttack.SetActive(true);
+            }
         }
 
         // Function gets executed by Player Input
         void OnSecondaryAttack()
         {
-            playerAttacks.AoeAttack(AoeAttack, transform.position);
+            if (activeAoeAttack == null)
+            {
+                animator.Play("Cast");
+                activeAoeAttack = Object.Instantiate(AoeAttack, transform.position, Quaternion.identity);
+            }
         }
 
         // Function get executed by Player Input
@@ -109,84 +89,6 @@ namespace Assets.Scripts.Characters.Player
             var relativePos = position - (Vector2)transform.position;
             var angle = Mathf.Atan2(relativePos.y, relativePos.x) * Mathf.Rad2Deg;
             return Quaternion.AngleAxis(angle, Vector3.forward);
-        }
-
-        public void TakeDamage(int damage)
-        {
-            var isCrit = RandomHelper.GetRandom(50);
-            if (isCrit)
-                damage *= 2;
-
-            currentHP -= damage;
-
-            CombatTextPopup.Damage(transform, 1.4f, damage, isCrit);
-
-            if (currentHP <= 0)
-                animator.Play("Death");
-            else
-                animator.Play("TakeHit");
-
-            healthGlobe.UpdateHealthGlobe((float)currentHP / healthPoints);
-        }
-
-        public void AggroMobs()
-        {
-            var aggroDistance = 10f;
-
-            var enemyObjects = FindObjectsOfType<BaseEnemy>();
-            foreach (var enemy in enemyObjects)
-            {
-                // Find enemies within 10f radius
-                if (Vector3.Distance(enemy.transform.position, transform.position) < aggroDistance)
-                {
-                    enemy.Move(transform.position);
-                }
-                else
-                {
-                    enemy.StopMovement();
-                }
-            }
-        }
-
-        public void AggroItems()
-        {
-            var consumeDistance = 0.5f;
-            var aggroDistance = 5f;
-
-
-            var gemObjects = FindObjectsOfType<Gem>();
-            foreach (var gem in gemObjects)
-            {
-                var playerPosition = transform.position + new Vector3(0, 1f);
-                // Find items within 5f radius
-                if (Vector3.Distance(gem.transform.position, playerPosition) < consumeDistance)
-                {
-                    Score.GetComponent<Score>().UpdateScore(gem.PickUp());
-                }
-                else if (Vector3.Distance(gem.transform.position, playerPosition) < aggroDistance)
-                {
-                    gem.Move(playerPosition);
-                }
-            }
-
-            var potionObjects = FindObjectsOfType<HealthPotion>();
-            foreach (var potion in potionObjects)
-            {
-                var playerPosition = transform.position + new Vector3(0, 1f);
-                // Find items within 5f radius
-                if (Vector3.Distance(potion.transform.position, playerPosition) < consumeDistance)
-                {
-                    var heal = potion.Consume();
-                    CombatTextPopup.Heal(transform, 1.4f, heal);
-                    currentHP += heal;
-
-                    healthGlobe.UpdateHealthGlobe((float)currentHP / healthPoints);
-                }
-                else if (Vector3.Distance(potion.transform.position, playerPosition) < aggroDistance)
-                {
-                    potion.Move(playerPosition);
-                }
-            }
         }
     }
 }
