+++ Assets/Scripts/Characters/Player/PlayerInventory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa17d5b8dcfaac2448974691735db806
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
