+++ Assets/Scripts/Characters/Player/PlayerStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 64b671fc5de11b14f9cfe4aa4393b16c
+guid: 4fc12ec1ae0be074c91fb1d4a5fa81e3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
